vendor: re-vendor agent to fix mount error
The agent gRPC protocol has changed meaning the order of the gRPC
`Storage` message members has changed. This manifested itself with the
following error:

```
"OCI runtime create failed: rpc error: code = Internal desc = Could not resolve symlink for source 9p: unknown."
```

This was caused by the agent attempting to mount a 9p FS with a blank
mountpoint - in fact due to the change in ordering, the mountpoint was
in the mount options!

Changes:

    9760c10 device: Add SCSI block device support
    15e851f storage: Rework the way Storage and Device are handled
    4683745 lint: Remove unnecessary conversions
    058182d lint: Fix vet error
    ebcb3af reaper: removed unused type
    902b82c device: Use standard functions to get device Major and Minor
    6e262da main: Display full stacktrace on internal error

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>